testing: move tt.go
