+++ internal/test/e2e/gouging_test.go
@@ -1,4 +1,4 @@
-package testing
+package e2e
 
 import (
 	"bytes"
